+++ src/middleware/MvcMiddleware.php
@@ -19,18 +19,14 @@ class MvcMiddleware extends AbstractMiddleware
      * @var ControllerFactoryInterface
      */
     private $controllerFactory;
-    private $templates;
 
-    public function __construct(ControllerFactoryInterface $controllerFactory, Templates $templates)
+    public function __construct(ControllerFactoryInterface $controllerFactory)
     {
         $this->controllerFactory = $controllerFactory;
-        $this->templates = $templates;
     }
 
     public function run($route, $response)
     {
-        $this->templates->prependPath('views/shared');
-        $this->templates->prependPath('views/layouts');
         $parameters = $route['parameters'] + Input::get() + Input::post();
         $controller = $this->controllerFactory->createController($parameters);
         return $this->controllerFactory->executeController($controller, $parameters);