+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-v0.6.0 - 2020-02-25
-===================
+# CHANGELOG
 
+## Unreleased
+
+### Added
+- Container builder now takes the `$namespace` parameter.
+
+### Changed
+- Routing config, dependency injection config, and middleware config are now read from a required `bootstrap` directory.
+- ORM subsystem is only initialized when a database configuration exists.
+
+
+## v0.6.0 - 2020-02-25
 First version with a CHANGELOG